Make 58 docstring headers say what their functions actually take
CLAUDE.md requires the function name, argument names and types in a docstring
header to match the `function` line character-for-character, and nothing checked
it.  Of 1897 docstring/signature pairs, 122 disagreed.

The interesting ones are not untidy, they are misleading:

  * PhotoIonization.Line and Einstein.Line documented crossSection::Float64 and
    photonRate::Float64 where both fields are EmProperty -- a reader would expect
    a scalar and get a two-gauge struct;
  * ImpactExcitationAutoion.displayPathways and its PhotoExcitationAutoion twin
    documented Array{...Line,1} where the code takes Array{...Pathway,1};
  * two Cascade routines documented Array{Configurations,1}, a type that does not
    exist in JAC at all;
  * Basics.determinePolarizationLambda documented the ABSTRACT
    Basics.AbstractPolarization where the method dispatches on the concrete
    Basics.LeftCircular -- the exact error CLAUDE.md names;
  * IsotopeShift.displayResults documented Array{Hfs.Outcome,1} for its own
    IsotopeShift.Outcome;
  * MultiPhotonTransition.computeLines still documented a `process` argument of
    type TwoPhotonAbsorptionMonochromatic, a scheme type retired in the rebuild.

58 headers in 41 files are corrected by layout-preserving token replacement.

WHAT IS DELIBERATELY LEFT.  The 41 argument-COUNT mismatches are excluded: adding
or removing an argument from a header needs judgement rather than a rewrite.  The
~22 qualification-only differences -- Array{Module.T,1} against Array{T,1} -- are
left as they are, because the type is the same and stripping the module prefix
from documentation makes it less clear, not more correct.  And 24 of the original
122 were my own detector mis-reading headers written as usage examples with
literal arguments; those are not defects.

ONE CASE IS LEFT FOR A DECISION rather than repaired.
InteractionStrength.MabEmissionJohnsony documents `a::Orbital, b::Orbital` while
the code reads `b::Orbital, a::Orbital`.  The arguments are positional, so a
caller who follows the docstring passes the a-orbital where the body calls it b;
for an asymmetric emission matrix element that is a physics error, and it is not
evident from the code which of the two carries the mistake.

The scanner is kept as work/diag-docstring-audit.py so the audit can be repeated.

Suite 43/43.

+++ src/module-BasicsAZ-inc-compute.jl
@@ -20,7 +20,7 @@ function Basics.compute(::AngularCoeffsEeRatip2013, csfa::CsfR, csfb::CsfR)
 end
 
 """
-`Basics.compute(::AngularCoeffs1pRatip2013, rank::Int64, csfa::CsfR, csfb::CsfR)`
+`Basics.compute(::AngularCoeffs1pRatip2013, rank, csfa::CsfR, csfb::CsfR)`
     ... to compute the angular coefficients in the decomposition of a (reduced) many-electron matrix element with a general
         single-particle operator of the given rank ``⟨csf_a ||O^rank|| csfb⟩ = \\sum_t  T(a_t, b_t) * R (a_t, b_t)``  by a call
         to the Fortran procedure `anco_calculate_csf_pair_1p` of the RATIP program; an `Array{AngularTcoeff,1}` is returned.
